Refactor: Move createCopyToClipboardFunction to its own file and update imports
- Created a new file for createCopyToClipboardFunction to improve code organization.
- Updated the import paths in SyntaxHighlighter and Code components to reflect the new structure.
- Removed the inline implementation of createCopyToClipboardFunction from SyntaxHighlighter.
